Isolating Quest Pro Parameters to the ExtTrackingInterface
special Quest Pro tracking parameters should be handled from the extTrackingInterface to make the system modular for VRCFT. Currently the extTrackingInterface is embedded in the VRCFT exe this needs to be removed from the main code.  Added Jaw Drop to work with MouthTowards for testing purposes instead compensating in code.